fix: metadata 미적용 문제 해결 및 text-kuit 색상 변경 문제 해결
