+++ src/app/layout.tsx
@@ -1,12 +1,13 @@
-'use client';
-
 import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
 import localFont from 'next/font/local';
 import './globals.css';
-import Background from '@/components/common/background/Background';
-import { usePathname } from 'next/navigation';
-import { metadata } from './metadata';
+import ClientLayout from '@/components/common/ClientLayout';
+
+export const metadata: Metadata = {
+  title: 'KUIT',
+  description:
+    'KUIT은 건국대학교 기획/개발 동아리입니다. 현재 4기에는 Android, Web, Server, PM, 디자이너 파트가 있으며, 10주간 스터디가 진행되고 방학에는 팀을 짜 프로젝트가 진행됩니다.',
+};
 
 const pretendard = localFont({
   src: '../../public/fonts/PretendardVariable.woff2',
@@ -20,18 +21,10 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const pathname = usePathname();
-
   return (
     <html lang="kr" className={`${pretendard.variable}`}>
       <body className={pretendard.className}>
-        <div className="relative min-h-screen">
-          <Background pathname={pathname} />
-          <div className="relative z-10">
-            {/* <Header /> */}
-            <main>{children}</main>
-          </div>
-        </div>
+        <ClientLayout>{children}</ClientLayout>
       </body>
     </html>
   );
