--- src/app/metadata.ts
@@ -1,7 +1,0 @@
-import { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  title: 'KUIT',
-  description:
-    'KUIT은 건국대학교 기획/개발 동아리입니다. 현재 4기에는 Android, Web, Server, PM 파트가 있으며, 10주간 스터디가 진행되고 방학에는 팀을 짜 프로젝트가 진행됩니다.',
-};
